Remove example functions now that we're handling CRDs
